requirements.txt
python-dotenv
aiohttp
asyncio
requests
langchain
chromadb
